edited regular expressions to be more inclusive and added test coverage in for deconfigged
added new test file
