+++ jdaviz/configs/imviz/tests/test_catalogs.py
@@ -30,8 +30,6 @@ from astropy.io import fits
 from astropy.nddata import NDData
 from astropy.table import Table, QTable
 
-from jdaviz.core.loaders.importers.catalog.catalog import CatalogImporter
-
 
 @pytest.mark.remote_data
 class TestCatalogs:
@@ -597,49 +595,3 @@ def test_select_catalog_table_rows(imviz_helper, image_2d_wcs):
     catalogs_plugin.select_none()
     assert len(plugin_table.selected_rows) == 0
 
-@pytest.mark.parametrize("coordinate_name", ['ra', 'dec'])
-def test_coord_column(deconfigged_helper,
-                      sky_coord_only_source_catalog,
-                      coordinate_name):
-    resolver = deconfigged_helper.loaders['object']._obj
-    importer = CatalogImporter(app=deconfigged_helper._app,
-                               resolver=resolver, parser=None,
-                               input=sky_coord_only_source_catalog)
-
-    variations_to_check = {(coordinate_name.upper(), coordinate_name.upper()),
-                           (coordinate_name + '_gaia', coordinate_name + '_gaia'),
-                           ('source' + coordinate_name, 'source' + coordinate_name),
-                           ('fluxradius', '---'),
-                           ('decrement', '---')}
-
-    for v in variations_to_check:
-        this_table = sky_coord_only_source_catalog.copy()
-        this_table.rename_column(coordinate_name, v[0])
-        importer._input = this_table
-        #print(v, this_table.keys())
-        assert importer._guess_coord_cols(coordinate_name)[0] == v[1]
-
-@pytest.mark.parametrize("pixel_name", ['x', 'y'])
-def test_pixel_column(deconfigged_helper,
-                      sky_coord_only_source_catalog,
-                      pixel_name):
-    resolver = deconfigged_helper.loaders['object']._obj
-    importer = CatalogImporter(app=deconfigged_helper._app,
-                               resolver=resolver, parser=None,
-                               input=sky_coord_only_source_catalog)
-
-    sky_coord_only_source_catalog.rename_column('ra', 'x')
-    sky_coord_only_source_catalog.rename_column('dec', 'y')
-
-    variations_to_check = {(pixel_name.upper(), pixel_name.upper()),
-                           (pixel_name + 'source', pixel_name + 'source'),
-                           (pixel_name + '_pix', pixel_name + '_pix'),
-                           ('pix_' + pixel_name, 'pix_' + pixel_name),
-                           ('galaxy', '---')
-                           }
-
-    for v in variations_to_check:
-        this_table = sky_coord_only_source_catalog.copy()
-        this_table.rename_column(pixel_name, v[0])
-        importer._input = this_table
-        assert importer._guess_coord_cols(pixel_name)[0] == v[1]
